Create modules files
